+++ http/unstable/cookie_map_test.ts
@@ -5,9 +5,9 @@ import {
   assertEquals,
   assertRejects,
   assertThrows,
-} from "../assert/mod.ts";
-import { FakeTime } from "../testing/time.ts";
-import { KeyStack } from "../crypto/keystack.ts";
+} from "../../assert/mod.ts";
+import { FakeTime } from "../../testing/time.ts";
+import { KeyStack } from "../../crypto/keystack.ts";
 
 import {
   CookieMap,
