+++ http/unstable/method_test.ts
@@ -1,6 +1,6 @@
 // Copyright 2018-2023 the Deno authors. All rights reserved. MIT license.
 
-import { assert, assertEquals } from "../assert/mod.ts";
+import { assert, assertEquals } from "../../assert/mod.ts";
 
 import { HTTP_METHODS, isHttpMethod } from "./method.ts";
